refactor: remove CmsModelFieldValidatorPlugin
