+++ packages/api-headless-cms/src/exports/api/cms/validation.ts
@@ -1,0 +1,2 @@
+export { CmsModelFieldValidator } from "~/features/validation/abstractions/CmsModelFieldValidator.js";
+export { CmsModelFieldValidatorRegistry } from "~/features/validation/abstractions/CmsModelFieldValidatorRegistry.js";
