+++ packages/api-headless-cms/src/features/validation/index.ts
@@ -1,2 +1,3 @@
-export { CmsModelFieldValidator, CmsModelFieldValidatorRegistry } from "./abstractions.js";
+export { CmsModelFieldValidator } from "./abstractions/CmsModelFieldValidator.js";
+export { CmsModelFieldValidatorRegistry } from "./abstractions/CmsModelFieldValidatorRegistry.js";
 export { ValidationFeature } from "./feature.js";
